refactor(RUN-952): split execution collaborators
